Server/Misc: Fix Arithmetic Overflow.

+++ src/server/game/Entities/Player/Player.cpp
@@ -6808,7 +6808,7 @@ void Player::SendActionButtons(uint32 state) const
         }
 
         buttonsTab[i].id = ab->GetAction();
-        buttonsTab[i].unk = uint32(ab->GetType()) << 56;
+        buttonsTab[i].unk = uint32(uint8(ab->GetType()) << 56);
 
     }
